+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `openstackmachinetemplate_controller` to remove finalizers from old&ununsed templates.
+- Add `OpenStackMachineTemplate` controller to remove finalizers from unused templates.
 
 ## [0.1.0] - 2022-02-15